mcp-server-with-postgresql-container/PostgresQL-Container/pyproject.toml
[project]
name = "postgresql-container"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "mcp>=1.13.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
